+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 ## New in git-machete 3.10.0
 
-- added: 3 git config keys `machete.github.{remote,organization,repository}` that enables `git machete github *` subcommands to work 
+- added: 3 git config keys `machete.github.{remote,organization,repository}` that enables `git machete github *` subcommands to work
   with custom GitHub URLs
 
 ## New in git-machete 3.9.1